@@ -1,6 +1,6 @@
 module github.com/pace/bricks
 
-go 1.16
+go 1.19
 
 replace github.com/adjust/rmq/v3 => github.com/daemonfire300/rmq/v3 v3.0.2
 
