@@ -2,7 +2,7 @@ package fueling
 
 import (
 	"context"
-	"io/ioutil"
+	"io"
 	"net/http/httptest"
 	"strings"
 	"testing"
@@ -53,7 +53,7 @@ func TestErrorReporting(t *testing.T) {
 
 	resp := rec.Result()
 	defer resp.Body.Close()
-	b, _ := ioutil.ReadAll(resp.Body)
+	b, _ := io.ReadAll(resp.Body)
 
 	require.Equalf(t, 422, resp.StatusCode, "expected 422 got: %s", string(b))
 	assert.Contains(t, string(b), `can't parse content: got value \"47.8\" expected type float32: Invalid type provided`)
